(WIP) Return errors and valid results from the same Entrypoint function
